Package name refactor. more robust permission screen
