+++ dev/app/src/main/AndroidManifest.xml
@@ -6,6 +6,8 @@
         android:name="android.hardware.telephony"
         android:required="true" />
     <uses-permission android:name="android.permission.READ_SMS"/>
+    <uses-permission android:name="android.permission.SEND_SMS"/>
+    <uses-permission android:name="android.permission.READ_CONTACTS"/>
 
     <application
         android:allowBackup="true"
@@ -18,7 +20,7 @@
         android:theme="@style/Theme.Dev"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.esms.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Dev">
