+++ dev/app/src/main/java/com/esms/models/Message.kt
@@ -1,0 +1,9 @@
+package com.esms.models
+
+import java.time.LocalTime
+
+data class Message(
+    val content: String,
+    val time: LocalTime,  //TODO: change to datetime
+    val recieved: Boolean //TODO: change to contact id
+)
