+++ dev/app/src/main/java/com/esms/theme/Type.kt
@@ -1,0 +1,8 @@
+package com.esms.theme
+
+import androidx.compose.material.Typography
+
+// Set of Material typography styles to start with
+val Typography = Typography(
+    // TODO: learn this
+)
