--- dev/app/src/main/java/com/example/dev/models/Message.kt
@@ -1,9 +1,0 @@
-package com.example.dev.models
-
-import java.time.LocalTime
-
-data class Message(
-    val content: String,
-    val time: LocalTime,  //TODO: change to datetime
-    val recieved: Boolean //TODO: change to contact id
-)
