--- dev/app/src/main/java/com/example/dev/theme/Type.kt
@@ -1,12 +1,0 @@
-package com.example.dev.theme
-
-import androidx.compose.material.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-
-// Set of Material typography styles to start with
-val Typography = Typography(
-    // TODO: learn this
-)
